requirements.txt
requests==2.27.1
python-dotenv==0.20.0
requests-oauthlib==1.3.1